Fingerprint/hardware_requirements.txt
# Hardware Scanner Test Dependencies
pyfingerprint>=1.5
pyserial>=3.5
